tools: allow ES6 getter @export + closure updates
This should fix a closure compiler error:

     @export only applies to symbols/properties defined
     in the global scope.

It is fixed by enabling export_local_property_definitions flag.
Unfortunately, it also results in NullPointerException
due to the new_type_inf flag. So, the latter had to go for tests.

The change also removes some of the suppress arguments
to the closure_js_library rules because they are unused anyway:
the compilation happens in closure_js_binary targets.

Lastly, upgraded rules_closure to the latest HEAD,
which supposedly makes a better use of the Closure Compiler.