Use image native width unless set
If an image does not have a defined width it is currently forced
to be 100% this can cause codelabs initially written in markdown
to have stretched images since there is no way to define image
size in markdown. From GDocs image width is automatically added.

This change allows both markdown and GDoc authors to get what
they expect when they publish a codelab from the corresponding
source.